requirements.txt
blinker==1.4
click==7.1.2
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-Mail==0.9.1
Flask-WTF==0.14.3
idna==3.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Werkzeug==1.0.1
WTForms==2.3.3
Flask-Login
Flask-Migrate
Flask-Script
Flask-SQLAlchemy
psycopg2